GitRepository: Fix reset(0) i.e. for reset to HEAD
[why]
It tries to reset to 'HEAD~0' which would in principle be fine, except
that it does not always pick HEAD (the reason for that is unclear for
me).

[how]
Add special rule that 'HEAD' is used as reference if we do not want to
go back but just reset to HEAD.

Prevent negative numbers as they do not make sense and we can avoid any
checks by using unsigneds.

[note]
In principle the interface with a number might be convenient at times
but is very limiting (i.e. unusable) generally. Better would be at least
an additional interface with a free reference string like done for
get_commit().

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>