test: Fix comments
Fix whitespaces:
Part of one comment has been created via `tree` but that seems to create
not only blanks (ascii 20) but also something else (ascii a0).
Convert the a0 to 20.

And one part of the comment looks like Doxygen but contains preformatted
code. Although the tests are currently not part of the Doxygen-ization
process we should not have broken output once it is.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>