Move all files produced by tests under unit_test_files/
[why]
To keep them in one place, and to be able to remove them with a single
command in the unit test main().

[how]
Introduce a function unit_test_folder() in main.h/cc that returns the
name of the folder, "unit_test_files", and use it for constructing file
names in the tests.

Signed-off-by: Lars Froehlich <lars.froehlich@desy.de>