Add proper Error class and enable Error tests
[why]
The libgit2 has a lot of error codes that we migth want to transport.
At the moment that is not possible and we can just generate some random
text messages that are hard to decifer for our users.

[how]
Use std::system_error which is able to transport error codes.

Define the error categories for the git related error codes, so that the
standard library can convert the number to text.

Enable the test_Error.cc and put something meaningful into it.

Add some extra constructors to the Error object that just take a string.
They were used before when a std::exception has been untilized which can
not hold error codes. The extra constructor uses in this case the error
code GIT_EUSER as reserved exactly for this purpose by libgit2.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>