wrapper: Implement missing branch_remote_name()
[why]
The function branch_remote_name() is already used but never defined.
This results in a library with missing symbols and when the function
depending on that symbol is called a crash.

[how]
There already was an almost complete but commented out implementation.
Add evaluation of the error code to that implementation and use the
Error object to raise an exception with code and explanantion.

Fixes: #19

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>