wrapper_functions: Simplify branch_remote_name() cleanup
[why]
The libgit2 buffer needs to be cleaned up in the end always. At the
moment we handle that manually.

[how]
Use a finalizer object that cleans up the allocated buffer when the
function ends in all cases.

It is always safe to call, because it will not free if it has not yet
been assigned/allocated.

[note]
I'm not sure about the variable name _ (underscore). It is in wide use
but in other places in libgit4cpp we call it 'cleanup'.
For me just an underscore is more obvious that it will not be used for
anything but RAII.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>