Use commented reinterpret_cast instead of C-style cast
[why]
libgit2 simulates inheritance by declaring an internal struct

struct git_commit {
	git_object object;

	git_vector parent_oids;
	git_oid tree_oid;

	git_signature *author;
	git_signature *committer;

	char *message;
	char *message_short;
};

whose declaration is not visible to library users. In C, it is legal to
cast a git_commit* to a git_object*.

[how]
We have no choice but do the same, but we spell out the reinterpret_cast
and add a comment why this mess is necessary. It is also quite likely
that this is undefined behavior for C++. Sigh.

Signed-off-by: Lars Froehlich <lars.froehlich@desy.de>