Debian: Change catch2 build dependency
This changes the build-depends clause from  "catch2 (>>3.4.0)" to
"catch2 (>=3.4.0) | libcatch2". Rationale (Fini):

"
Hmm, we have a problem here, out there is just this one package that
is an in-between one:

3.4.0-1build1: Only on Ubuntu, and there on 24.04 and 24.10.

    The header only catch2 packages were called catch2
    The library v3 version is destined to be called libcatch2
    There is a transitional package transmogrifying catch2 to libcatch2

On Ubuntu systems after 24.10 it is possible to install libcatch2 and
not have catch2 installed.

On Debian we do not have this break, and Catch2v3 is always available
as libcatch2.
"

Co-authored-by: Fini Jastrow <ulf.fini.jastrow@desy.de>
Signed-off-by: Lars Froehlich <lars.froehlich@desy.de>